Move from CentOS to Debian. **BEWARE! INTRODUCING BREAKING CHANGES!**
After ff0b661e112c00b8b98ca3051e68232b45b97aba, it failed with this:

    [ALERT] 252/070959 (10) : parsing [/etc/haproxy/50-https.cfg:21] : unknown keyword 'ssl-default-bind-options' in 'global' section
    [ALERT] 252/070959 (10) : parsing [/etc/haproxy/50-https.cfg:24] : unknown keyword 'ssl-default-server-options' in 'global' section
    [ALERT] 252/070959 (10) : Error(s) found in configuration file : /etc/haproxy/50-https.cfg

That seems to be a bug in CentOS because the Debian version behaves as expected.
That's the rationale of this change.

To follow the Debian standard, certs now go to `/etc/ssl/private/`. If you used
environment variables to put the certs, then it will behave as usual; but if
you mounted volumes or added files, you must change the path.